pyproject.toml
[tool.poetry]
name = "debutizer"
version = "0.1.0"
description = "A framework for creating APT packages"
authors = ["Tyler Compton <xaviosx@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
python-debian = "^0.1.40"
xdg = "^5.1.1"

[tool.poetry.dev-dependencies]
pre-commit = "^2.15.0"
pytest = "^6.2.5"
black = "^21.9b0"
isort = "^5.9.3"
mypy = "^0.910"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
src_paths = "debutizer"
